Add neon_fcmla and neon_fcadd as neon_type instructions.
[AArch64 folks CC'd fyi as this is common between both backends.]

Hi,

The design in the backend used to be that advanced simd types are
generally added to is_neon_type in the backend. It appears that
neon_fcmla and neon_fcadd aren't added in as  neon_type instructions.

Applying this to the tree later this week after having built armhf and
a bootstrap and test run on aarch64-linux-gnu.

Thanks,
Ramana
commit 7dd15fae0ac1455f5818a1fc0078e35d85e1e250
Author: Ramana Radhakrishnan <ramana.gcc@gmail.com>
Date:   Wed Nov 16 10:32:04 2022 +0000

    [Patch Arm] Add neon_fcadd and neon_fcmla to is_neon_type.

    Appears to have been an oversight.

    gcc/
            * config/arm/types.md: Update comment.
            (is_neon_type): Add neon_fcmla, neon_fcadd.

    Signed-off-by: Ramana Radhakrishnan <ramana.gcc@gmail.com>