ada: Tweak error messages on misplaced with keywords
Before this patch, with clauses placed in declarative sections were
interpreted by the compiler as incorrect aspect specifications, which
led to confusing error messages.

This patch makes it so more syntax errors involving the with keyword
are diagnosed as intended with clauses instead of aspect
specifications.

gcc/ada/

	* par-ch3.adb (P_Declarative_Item): Tweak handling of with keyword.

Tested on x86_64-pc-linux-gnu, committed on master.