tree-optimization/106912 - IPA profile and pure/const
IPA profile instrumentation tries to clear the pure and const
flags of functions but that's quite hopeless in particular for
const since that attribute prevails on the type and thus on each
call to the function leading to inconsistencies in the IL and
eventual checking ICEs.  There's no good reason to do this and
it wouldn't fixup any indirect calls so just don't.  No other
instrumentation GCC does bothers about this.

Bootstrap and regtest pending on x86_64-unknown-linux-gnu, OK?

Thanks,
Richard.

	PR tree-optimization/106912
	* tree-profile.cc (tree_profiling): Do not clear pure/const
	flags.

	* gcc.dg/pr106912.c: New testcase.