Fix unrecognizable insn due to illegal immediate_operand (const_int 255) of QImode.
For __builtin_ia32_vec_set_v16qi (a, -1, 2) with
!flag_signed_char. it's transformed to
__builtin_ia32_vec_set_v16qi (_4, 255, 2) in the gimple,
and expanded to (const_int 255) in the rtl. But for immediate_operand,
it expects (const_int 255) to be signed extended to
(const_int -1). The mismatch caused an unrecognizable insn error.

expand_expr_real_1 generates (const_int 255) without considering the target mode.
I guess it's on purpose, so I'll leave that alone and only change the expander
in the backend. After applying convert_modes to (const_int 255),
it's transformed to (const_int -1) which fix the issue.

Bootstrapped and regtested x86_64-pc-linux-gnu{-m32,}.
Ok for trunk(and backport to GCC-10/11/12 release branches)?

gcc/ChangeLog:

	PR target/107863
	* config/i386/i386-expand.cc (ix86_expand_vec_set_builtin):
	Convert op1 to target mode whenever mode mismatch.

gcc/testsuite/ChangeLog:

	* gcc.target/i386/pr107863.c: New test.