c++: TYPENAME_TYPE lookup ignoring non-types [PR107773]
[temp.res.general]/3 says, in a note, "the usual qualified name lookup
([basic.lookup.qual]) applies even in the presence of typename".  Thus
when resolving a TYPENAME_TYPE, it seems we shouldn't be looking past
non-type members.

This patch fixes this by passing want_type=false instead of =true during
the member lookup from make_typename_type.  An old nearby comment
mentions that we want to continue to set want_type=true when resolving a
nested typename type, but it appears that the nested case is handled by
resolve_typename_type instead (which passes want_type=true appropriately).

In passing, use lookup_member instead of lookup_field so that we give a
better diagnostic when a member function is found, and generalize the T
format specifier to D in the diagnostic.

Bootstrapped and regtested on x86_64-pc-linux-gnu, does this look OK for
trunk?

	PR c++/107773

gcc/cp/ChangeLog:

	* decl.cc (make_typename_type): Use lookup_member instead of
	lookup_field.  Pass want_type=false instead of =true.  Use D
	instead of T format specifier.
	* search.cc (lookup_member): Document default argument.

gcc/testsuite/ChangeLog:

	* g++.dg/template/typename24.C: New test.
	* g++.dg/template/typename25.C: New test.