amdgcn: Support AMD-specific 'isa' traits in OpenMP context selectors
Hi all,

This patch adds support for 'gfx803' as an alias for 'fiji' in OpenMP
context selectors, so as to be consistent with LLVM. It also adds test
cases checking all supported AMD ISAs are properly recognised when used
in a 'declare variant' construct.

Is it OK for mainline?

Thanks,