IPA: do not release body if still needed
Hi.

Noticed during building of libbackend.a with the LTO partial linking.

The function release_body is called even if clone_of is a clone
of a another function and thus it shares tree declaration. We should
preserve it in that situation.

Patch can bootstrap on x86_64-linux-gnu and survives regression tests.

Ready to be installed?
Thanks,
Martin

	PR ipa/107944

gcc/ChangeLog:

	* cgraph.cc (cgraph_node::remove): Do not release body
	if a node is clone of another node.