libgcc: Fix uninitialized RA signing on AArch64 [PR107678]
A recent change only initializes the regs.how[] during Dwarf unwinding
which resulted in an uninitialized offset used in return address signing
and random failures during unwinding.  The fix is to use REG_SAVED_OFFSET
as the state where the return address signing bit is valid, and if the
state is REG_UNSAVED, initialize it to 0.

Passes bootstrap & regress, OK for commit?

libgcc/
        PR target/107678
        * unwind-dw2.c (execute_cfa_program): Initialize offset of
        DWARF_REGNUM_AARCH64_RA_STATE if in REG_UNSAVED state.
        * config/aarch64/aarch64-unwind.h (aarch64_frob_update_contex):
        Check state is REG_SAVED_OFFSET before using offset for RA state.