-fno-short-enums for auto-init-[12].c
On arm-eabi, and possibly on other platforms, -fshort-enums is enabled
by default, which breaks some tests' expectations as to enum sizes
with DEFERRED_INIT.  Disable short enums so that the expectations are
met.

Regstraped on x86_64-linux-gnu, also tested with crosses to riscv64-elf
and arm-eabi.  Ok to install?

for  gcc/testsuite/ChangeLog

	* c-c++-common/auto-init-1.c: Add -fno-short-enums.
	* c-c++-common/auto-init-2.c: Likewise.
	* gcc.dg/debug/btf/btf-enum-1.c: Likewise.