disable tree-dce for test
CD-DCE introduces blocks to share common PHI nodes, which replaces a
backwards branch that used to prevent the thumb2 jump table shortening
that PR42093 tested for.  In order to keep on testing that the
backward branch prevents the jumptable shortening, disable tree-dce.

Regstraped on x86_64-linux-gnu, also tested with crosses to riscv64-elf
and arm-eabi.  Ok to install?

for  gcc/testsuite/ChangeLog

	PR target/42093
	* gcc.target/arm/pr42093.c: Disable tree-dce.