xfail fp-uint64-convert-double tests
The FP emulation on ARM doesn't take rounding modes into account.  The
tests require hard_float, but that only tests for calls when adding
doubles.  There are arm targets that support hardware adds, but that
emulate conversions.

Regstraped on x86_64-linux-gnu, also tested with crosses to riscv64-elf
and arm-eabi.  Ok to install?

PS: I've pondered changing hard_float instead, but decided that could
impact other tests that rely on its current behavior.  Then I considered
adding a variant of hard_float that tested for conversions, but thought
that it was overkill.

for  gcc/testsuite/ChangeLog

	* gcc.dg/torture/fp-uint64-convert-double-1.c: Expect fail on
	arm-*-eabi*.
	* gcc.dg/torture/fp-uint64-convert-double-2.c: Likewise.